app 분리 및 models.py 수정
