+++ backend/settings.py
@@ -38,7 +38,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'bookg',
+    'user',
+    'book',
+    'page',
     'drf_yasg',
 ]
