+++ docker-compose.yml
@@ -7,12 +7,21 @@ services:
       MYSQL_DATABASE: test
     ports:
       -  "3306:3306"
+    healthcheck:
+      test: [ "CMD", "mysqladmin", "ping", "-h", "localhost" ]
+      interval: 10s
+      retries: 5
+      timeout: 5s
 
   web:
     build: .
+    depends_on:
+      db:
+        condition: service_healthy
     ports:
       - "8000:8000"
     volumes:
       - .:/app
-    depends_on:
-      - db
+    restart: always
+    environment:
+      WAIT_HOSTS: db:3306
