+++ user/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BookgConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'bookg'
+    name = 'user'
